v3: Create 'roles' view to back Role model, instead of dataset
We had a dataset representing the union of the 7 roles tables that was
backing the Role Sequel model. This change adds a migration to create a
view that represents the same union, and switches the model to be backed
by the view instead.

This solves a problem introduced as part of #164176828, where the
PollableJob model was trying to run a query against a table called
`roles` to check if a role existed. Since that table didn't exist, it
was failing. Now, there is a view with that name, so those queries work.

[Finishes #164176828]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>